Added reach-while-avoid certificate (identical to RWS but framed with unsafe set). RWS is not currently supported.
Added benchmark for this, but doesn't work well

Fixed some bugs in RWA. Works much better now. Benchmark updated

Added unified function to plot lie derivative, and helper to plot all of phase plane, cert and lie with sets for a learner.

Also there's a bug with the RWA. With the control case study, it is certified even though it is clearly not correct (see plots).

Now updated RWS to work. The control example for this is kinda brittle, we need to encourage stability better I think.

The RWA is theoretically incorrect - to be discussed

CUrrently no benchmarks for RSWA and RSWS. In theory these could work but it's dependent on a good choice for the parameter \beta, which is unclear